# openapi.yaml
openapi: 3.0.0
info:
  title: Obscuron Chat Core API
  description: Obscuron Chat is a web-based encrypted chat app
  version: 1.0.0
servers:
  - url: https://api.obscuron.chat
    description: Production Server
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /messages/{peer}:
    get:
      summary: Fetch encrypted conversation history
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: peer
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9_]{3,32}$'
          description: Peer username; must not be the authenticated user.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Maximum number of encrypted messages to return.
        - in: query
          name: cursor
          schema:
            type: string
          description: Opaque pagination cursor from the previous response.
      responses:
        '200':
          description: Encrypted message history in descending server-received order.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid peer, limit, or cursor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Peer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        data:
          type: object
      required: [status, message, data]
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            status:
              type: string
              enum: [error]
          required: [status]
...